# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Danceseek API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_me_get",
        "parameters": [
          {
            "name": "ds_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ds Session"
            }
          },
          {
            "name": "ds_pending",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ds Pending"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/lastfm/callback": {
      "get": {
        "summary": "Lastfm Callback",
        "description": "Trade the approved token for a session and sign the user in.\n\nUses the token Last.fm hands back, falling back to the one we stashed at\nthe start of the flow — a user who returns by any route still lands signed in.",
        "operationId": "lastfm_callback_auth_lastfm_callback_get",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "ds_pending",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ds Pending"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/lastfm/complete": {
      "post": {
        "summary": "Lastfm Complete",
        "description": "Finish a connection whose callback never arrived.\n\nLast.fm only redirects back when the API account has a Callback URL\nregistered; without one it shows its own success page and leaves the user\nto navigate back. The approved token is still redeemable, so this closes\nthe loop from the app side.",
        "operationId": "lastfm_complete_auth_lastfm_complete_post",
        "parameters": [
          {
            "name": "ds_pending",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ds Pending"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/setlists": {
      "get": {
        "summary": "List Setlists",
        "description": "Browse rows. Filters repeat to multi-select: `?dj=A&dj=B` matches either.\n\nThere's no total count — callers asking for `limit + 1` and checking whether\nthey got it is enough to drive a \"load more\", and avoids a second query.",
        "operationId": "list_setlists_setlists_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "dj",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dj"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Genre"
            }
          },
          {
            "name": "event",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SetlistSummary"
                  },
                  "title": "Response List Setlists Setlists Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/setlists/{setlist_id}": {
      "get": {
        "summary": "Get Setlist",
        "operationId": "get_setlist_setlists__setlist_id__get",
        "parameters": [
          {
            "name": "setlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Setlist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetlistDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/setlists/{setlist_id}/cues": {
      "get": {
        "summary": "Setlist Cues",
        "description": "Cue windows: which track is playing when, and under what name.\n\nThe client reports the player's duration; everything else is derived here so\nthe highlight and the scrobbler can never disagree about what's playing.\nEligibility reflects the signed-in user's settings, so the tracklist shows\nexactly what would be scrobbled.",
        "operationId": "setlist_cues_setlists__setlist_id__cues_get",
        "parameters": [
          {
            "name": "setlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Setlist Id"
            }
          },
          {
            "name": "duration",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Recording length in seconds, when known",
              "title": "Duration"
            },
            "description": "Recording length in seconds, when known"
          },
          {
            "name": "ds_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ds Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindowSet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scrobble/now-playing": {
      "post": {
        "summary": "Now Playing",
        "description": "Flag what's playing. Not a play — no dedupe, nothing recorded.",
        "operationId": "now_playing_scrobble_now_playing_post",
        "parameters": [
          {
            "name": "ds_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ds Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrobbleTarget"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrobbleResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Me": {
        "properties": {
          "lastfm_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastfm Username"
          },
          "connected": {
            "type": "boolean",
            "title": "Connected",
            "default": false
          },
          "pending": {
            "type": "boolean",
            "title": "Pending",
            "default": false
          }
        },
        "type": "object",
        "title": "Me"
      },
...
      "SetlistDetail": {
        "properties": {
          "setlist": {
            "$ref": "#/components/schemas/Setlist"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "coverage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "setlist",
          "status",
          "coverage",
          "resolved_at",
          "created_at"
        ],
        "title": "SetlistDetail",
        "description": "Full setlist plus its resolution lifecycle metadata."
      },
...
    }
  }
}